Backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy==2.0.35
psycopg[binary,pool]==3.2.3
alembic==1.13.3
pydantic[email]==2.9.0
pydantic-settings==2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-dotenv==1.0.1
aiosmtplib==3.0.1
openpyxl==3.1.5
python-multipart==0.0.9
greenlet==3.1.1
apscheduler==3.10.4
